refactor wip, now we have one factory that return the api, not two factories..., load function changed - only embedOptions are handled inside ekoDelivery.load, the rest (events, ea, etc...) is in ekoPlayer load
