+++ src/lib/ekoPlatform.js
@@ -1,0 +1,7 @@
+function isEkoDomain(domain) {
+    return /https?:\/\/(.*?\.)?eko.com/.test(domain);
+}
+
+export default {
+    isEkoDomain
+};
